Write a .changes file
Modern distributions use source_date_epoch_from_changelog to supply a
stable SOURCE_DATE_EPOCH to build systems for reproducible builds. This
however requires at least a minimal changelog to be present.

We'll want to have one in the long run anyway, so may as well get
started on that.